Improve the nick API.
Here is a better API: Rather than register/unregister, support a
swap(From, To) function. This way, we specially handle a swap from "*"
to a nick as the first kind of swap and then we handle swaps of nicks
afterwards. With this, there is a simple way to know if there is an
old nick we have to take care of.